refactor: 많은 양의 API에 대응하기 위한 명시적 구조 개선 작업
