+++ Cheffi/Network/EndPoint/RestRouter.swift
@@ -1,0 +1,14 @@
+//
+//  RestRouter.swift
+//  Cheffi
+//
+//  Created by 이서준 on 6/7/24.
+//
+
+import Foundation
+
+enum RestRouter {
+    case testUpload
+    case testSessionIssue
+    case testAuth(String)
+}
